backend/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.1.8
fastapi==0.115.12
graphql-core==3.2.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
strawberry-graphql==0.262.5
typing_extensions==4.12.2
uvicorn==0.34.0
python-jose==3.4.0
passlib==1.7.4
pyjwt==2.6.0
razorpay==1.3.0
